requirements.txt
streamlit
pinecone
sentence-transformers
google-generativeai
python-dotenv
huggingface_hub
requests